+++ CHANGELOG.rst
@@ -175,6 +175,8 @@ Changed
 - Replace the ``extract`` methods in ``walkers`` to be private.
 - Replace the ``graph.py`` file for ``graphs/kg.py``.
 - Replace the ``rdf2vec`` module for ``pyrdf2vec``.
+- Replace the ``sample_neighbor`` method of the ``sampler`` class by
+  ``sample_hop``.
 - Replace the imec licence for an MIT licence.
 - Remove ``graph`` hyper-parameter in the ``transform`` method of the ``RDF2VecTransformer`` class.
 - Remove hyper-parameters of ``RDF2VecTransformer`` for ``embedder`` and ``walkers`` ones.